requirements.txt
PyPDF2
langchain_openai
langchain_core
streamlit
langchain_community
python-dotenv
pypdf
langserve
fastapi
uvicorn
sse_starlette